+++ bin/tasks/rename.sh
@@ -9,6 +9,9 @@ sed -i '' -e "s;${BLOCK_NAME};${BLOCK_NAME_TO_RENAME_TO};" ${PROJECT_THEME_PATH}
 sed -i '' -e "s;${BLOCK_NAME};${BLOCK_NAME_TO_RENAME_TO};" ${PROJECT_THEME_PATH}/functions.php
 
 # Renaming block UI label
+# NB! This replaces ALL titles in the functions.php, it should replace only the
+# current block title.
+# TODO: Figure out how to replace only the current block title.
 sed -i '' -e "s/'title' \=\> '.*',/'title' \=\> '${BLOCK_UI_TITLE_TO_RENAME_TO}',/" ${PROJECT_THEME_PATH}/functions.php
 
 # Renaming block UI name in ACF